split production to separate file
